requirements.txt
future==0.16.0
kafka-python==1.3.3
parsedatetime==2.3
psycopg2==2.7.1
pytz==2017.2
